+++ src/app/common/components/dataset-list-component/dataset-list.component.ts
@@ -7,6 +7,7 @@
 
 import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { DatasetSearchOverviewFragment } from "src/app/api/kamu.graphql.interface";
+import { SearchMode } from "src/app/interface/search.interface";
 
 @Component({
     selector: "app-dataset-list",
@@ -16,13 +17,18 @@ import { DatasetSearchOverviewFragment } from "src/app/api/kamu.graphql.interfac
 })
 export class DatasetListComponent {
     @Input({ required: true }) public dataSource: DatasetSearchOverviewFragment[];
-    @Input({ required: true }) public totalCount = 0;
+    @Input({ required: true }) public totalCount: number;
     @Input({ required: true }) public resultUnitText: string;
-    @Input({ required: true }) public hasResultQuantity?: boolean = false;
-    @Input({ required: true }) public isClickableRow?: boolean = false;
+    @Input({ required: true }) public isClickableRow: boolean;
+    @Input({ required: true }) public searchMode: SearchMode;
     @Input({ required: true }) public sortOptions: {
         value: string;
         label: string;
         active: boolean;
     }[];
+    @Input() public hasResultQuantity?: boolean = false;
+
+    public get isSemanticSearch(): boolean {
+        return this.searchMode === SearchMode.SEMANTIC_SEARCH;
+    }
 }